Fixed bug where random questions would load regardless of category chosen by the player
Forgot to set the id of the button element to the id of the category which caused a bug that wouldn't load the proper quiz for the chosen category. Before, it would just give you random questions regardless of the category chosen. Now it gives you the proper questions based on the category chosen by the player.